Change all occurrences of _rect and _bounds Rectangles to _fillBounds and _lineBounds BoundingBoxes
Part 4 of the grand "we change all the rectangles to internal representations" refactoring.

Also adds verification code to BoundingBox, ensuring that all properties are integers and (currently disabled because of bugs in using code) that *Max > *Min.

+++ src/player/remotingGfx.ts
@@ -124,10 +124,10 @@ module Shumway.Remoting.GFX {
     private _readRectangle(): Rectangle {
       var input = this.input;
       return new Rectangle (
-        input.readFloat() / 20,
-        input.readFloat() / 20,
-        input.readFloat() / 20,
-        input.readFloat() / 20
+        input.readInt() / 20,
+        input.readInt() / 20,
+        input.readInt() / 20,
+        input.readInt() / 20
       );
     }
